# backend/swagger/e-hotel.yaml
openapi: 3.0.0
info:
  title: E-Hotel API
  version: 1.0.0
  description: API for managing resort hotels, room bookings, and staff authentication.
paths:
...
  /api/reports/{id}/download:
    get:
      summary: Download a performance report as a PDF file
      tags: [Reports]
      parameters:
        - name: id
          in: path
          required: true
          description: The report id
          schema:
            type: integer
      responses:
        '200':
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '404':
          description: Report not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...